refactor(app): thin out AppSidebar and fix render-time atom reads
Architecture cleanup from plan 005 (P3.2, P3.5):

- Replace imperative navigation.resolveAtoms() reads during render with
  useAtomValue subscriptions in PageEditor, PageAsset, and the note
  breadcrumb; the workspace-not-found views now re-render when the
  routed workspace changes instead of showing a stale name.
- Extract the sidebar file-tree context-menu actions into a
  useSidebarFileActions hook and the footer dropdown into
  SidebarFooterMenu, leaving AppSidebar as a thin composition layer
  (554 -> 150 lines) with no behavior change.

Verified via typecheck, unit suites, and the app-shell/file-explorer
Playwright suites.

Also verified while auditing plan 005: P2.1 (types package importing
core services), P2.2 (package-private src imports), and P3.1
(duplicate component trees) are already resolved on main, and the
FileSystemService storage wiring (P2.3) is now a config thunk asserted
at instantiate() rather than post-construction mutation; plan doc
update follows in the final commit.
